Refactor common helper for matching MDC version.
